Ajustando Middlewares

+++ src/controllers/ProdutosController.js
@@ -48,17 +48,30 @@ module.exports = {
 
             }
             errors.forEach(erro => 
-                errosFormatados[erro.param] = erro.msg               
+            errosFormatados[erro.param] = erro.msg               
             );
             console.log("Aqui 2" , errosFormatados)
             return res.render('cadastroProduto' , { pageName: 'cadastroProduto', js: 'montarCarrinho', errors: errosFormatados, produtos: null });
         }
-        ProdutosModel.createOne(req)
+            ProdutosModel.createOne(req)
             res.send(`O produto ${req.body.nome} foi criado com sucesso`)
     }, 
     //procurar produto
     buscaProduto: (req, res) => {
+        const { errors } = validationResult(req);
+        console.log("errors", errors)
+
+        if (errors.length){
+            const errosFormatados = {
+
+            }
+            errors.forEach(erro => 
+            errosFormatados[erro.param] = erro.msg               
+            );
+            return res.render('cadastroProduto' , { pageName: 'cadastroProduto', js: 'montarCarrinho', errors: errosFormatados, produtos: null });
+        }
         res.send(ProdutosModel.findOne(req,res))
+        
 
         // let filme = ProdutosModel.findOne(req);
         // return res.render('produtos', { pageName: 'produtos', js:' ', filme });
@@ -69,12 +82,40 @@ module.exports = {
     },
     //deletar produto
     deletaProduto: (req, res) => {
+        const { errors } = validationResult(req);
+        console.log("errors", errors)
+
+        if (errors.length){
+            console.log("Aqui")
+            const errosFormatados = {
+
+            }
+            errors.forEach(erro => 
+            errosFormatados[erro.param] = erro.msg               
+            );
+            console.log("Aqui 2" , errosFormatados)
+            return res.render('cadastroProduto' , { pageName: 'cadastroProduto', js: 'montarCarrinho', errors: errosFormatados, produtos: null });
+        }
         
         ProdutosModel.deleteOne(req)
         res.send(`O produto de id ${req.body.id} foi deletado com sucesso`)
     },
     // atualização de produto
     atualizaProduto: (req, res) => {
+        const { errors } = validationResult(req);
+        console.log("errors", errors)
+
+        if (errors.length){
+            console.log("Aqui")
+            const errosFormatados = {
+
+            }
+            errors.forEach(erro => 
+            errosFormatados[erro.param] = erro.msg               
+            );
+            console.log("Aqui 2" , errosFormatados)
+            return res.render('cadastroProduto' , { pageName: 'cadastroProduto', js: 'montarCarrinho', errors: errosFormatados, produtos: null });
+        }
         ProdutosModel.updateOne(req)
         res.send(`O produto de id ${req.body.id} foi atualizado com sucesso`)
     }
